D-MASH Alpha-1 TACT/client/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
websockets==12.0
aiosqlite==0.19.0
PyNaCl==1.5.0
pydantic==2.6.0
python-multipart==0.0.9